requirements.txt
python-telegram-bot==20.7
openai==1.12.0
google-api-python-client==2.116.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
